Dir: 404페이지 작업 경로 변경(#41) 기존의 404폴더 삭제하고 NotFound 폴더로 이동하여 작업
